Hoan thien frontend
