+++ src/components/pages/managerPage/DesktopSidebar.tsx
@@ -102,7 +102,7 @@ const DesktopSidebar = ({ navState, setNavState }: Props) => {
               [VaitroNguoiDung.KeToan].includes(user.vaiTroNguoiDung) &&
               checkMatchBaseRoutes(location.pathname) && (
                 <button
-                  onClick={() => navigate("/manager")}
+                  onClick={() => navigate("/account/show")}
                   className={classNames(
                     "text-gray-700 hover:text-gray-900 hover:bg-gray-50 group flex items-center px-2 py-2 text-sm font-medium rounded-md w-full"
                   )}
